Fix generic lerp never reaching target
This will not pose a problem in practice most of the time, because float
and double have specific overrides that do work. The generic
implementation however always returned a0, which means that the promise
of a1 being returned when t = 1 never holds true.

If lerping is not supported for a type, the generic implementation
should probably always return a1 instead, unless t = 0, immediately
reaching the destination. Otherwise, interpolating values gradually over
the network will keep trying to get closer to the destination value in
YourGeneratedNetworkObject.InterpolateUpdate,
but will never reach its destination.