Fix interpolation starting from incorrect source when disowned
The interpolation target is used on clients in combination with the
source to interpolate from the old value to the new. The owner, however,
does not need it, as it is the authority on what the values are in the
first place.

When the owner later on changes to another object, the original owner
will also start interpolating, as expected, and will also call
SnapInterpolations to ensure all interpolation immediately resolves to
its target.

This is fine, but the owner never kept track of the target interpolation
in the first place, as it's only read by others. This causes the owner
to snap to the values it had from before it became owner (or the default
values if it was the first owner).

This fixes that by also keeping the target around on the owner, where it
can then be  used in that corner case.