Implement TAP-based request-response flow on top of RPCs
This introduces requests with responses, which are just RPCs that can
send back a value. Requests are also just layered on top of RPCs and
don't introduce any new special treatment, they are just some
convenience sugar on top of RPCs where you want to send back a response.

Sending a request will return a Task and follow TAP [1], so you can
easily await the response to your request from the receiver.

[1] https://docs.microsoft.com/en-us/dotnet/standard/asynchronous-programming-patterns/task-based-asynchronous-pattern-tap